+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Improvement:
 - support resolve macro on ws
 - [#1146](https://github.com/elysiajs/elysia/pull/1146) add support to return web API's File from handler
 - [#1165](https://github.com/elysiajs/elysia/pull/1165) skip non-numeric status codes in response schema validation
+- [#1177](https://github.com/elysiajs/elysia/issues/1177) cookie does not sign when an error is thrown
 
 Bug fix:
 - `Response` returned from `onError` is using octet stream